server/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
coverage==7.9.2
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-unfold==0.59.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
pytest==8.4.1
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
